+++ src/shared/AppLogo/AppLogo.module.scss
@@ -1,0 +1,14 @@
+.logo {
+  margin-right: 30px;
+}
+
+.text {
+  color: rgb(30, 41, 59);
+  text-transform: uppercase;
+
+  font-size: 55px;
+  font-weight: 600;
+  line-height: 36px;
+  letter-spacing: 0.1px;
+  text-align: center;
+}
