+++ src/shared/AppLogo/AppLogo.tsx
@@ -1,4 +1,3 @@
-import cn from 'clsx';
 import React, { memo } from 'react';
 import s from './AppLogo.module.scss';
 
@@ -8,10 +7,9 @@ interface AppLogoProps {
 
 export const AppLogo = memo(({ className }: AppLogoProps) => {
   return (
-    <a className={cn(s.outer, className)} href="/">
-      <span className={s.title}>Приложение для.</span>
-      <span className={s.subtitle}>ведения бюджета</span>
-    </a>
+    <div className={s.logo}>
+      <span className={s.text}>SINET</span>
+    </div>
   );
 });
