+++ src/shared/AppLogo/index.ts
