+++ src/shared/Header/Header.module.scss
@@ -1,30 +1,28 @@
-.outer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: var(--header-height);
-    border-bottom: 1px solid var(--color-gray-3);
-  }
-  
-  .container {
-    position: relative;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    @include container();
-  }
-  
-  .logo {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-  }
-  
-  .right {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    margin-left: auto;
-  }
-  
+.header {
+  width: 100%;
+  box-sizing: border-box;
+  border-bottom: 1px solid rgb(147, 197, 253);
+  background-color: rgb(255, 255, 255);
+  padding: 19px 32px;
+
+  display: flex;
+  align-items: center;
+  word-wrap: break-word;
+}
+
+.wrapper {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  width: 100%;
+
+}
+
+.inform {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  align-items: center;
+
+}
