+++ src/shared/layouts/BaseLayout/BaseLayout.module.scss
@@ -1,0 +1,10 @@
+.layout {
+  display: flex;
+  flex-direction: column;
+  padding-top: auto;
+}
+
+.headerLayout{
+  position: sticky;
+  top: 0;
+}
