--- src/shared/ui/AppLogo/AppLogo.module.scss
@@ -1,20 +1,0 @@
-.outer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  .title {
-    margin-bottom: 0.2em;
-    font-weight: 600;
-    font-size: 28px;
-    line-height: 1;
-  }
-  
-  .subtitle {
-    font-weight: 500;
-    font-size: 12px;
-    letter-spacing: 0.2em;
-    text-transform: uppercase;
-  }
-
