--- src/shared/ui/layouts/BaseLayout/BaseLayout.module.scss
@@ -1,9 +1,0 @@
-.header {
-  position: sticky;
-  top: 0;
-}
-
-.main {
-  @include container();
-  padding: 24px;
-}
